+++ tests/client_tests.ts
@@ -1,4 +1,4 @@
-import * as Types from '../src/types'
+import MeiliSearch, * as Types from '../src/types'
 import {
   clearAllIndexes,
   config,
@@ -7,7 +7,6 @@ import {
   publicClient,
   anonymousClient,
   PUBLIC_KEY,
-  PRIVATE_KEY,
 } from './meilisearch-test-utils'
 
 const uidNoPrimaryKey = {
@@ -17,10 +16,6 @@ const uidAndPrimaryKey = {
   uid: 'movies_test2',
   primaryKey: 'id',
 }
-
-beforeAll(() => {
-  return clearAllIndexes(config)
-})
 
 afterAll(() => {
   return clearAllIndexes(config)
@@ -42,7 +37,7 @@ describe.each([
       await expect(client.listIndexes()).resolves.toHaveLength(0)
     })
     test(`${permission} key: create with no primary key`, async () => {
-      await client.createIndex(uidNoPrimaryKey).then((response) => {
+      await client.createIndex(uidNoPrimaryKey.uid).then((response) => {
         expect(response).toHaveProperty('uid', uidNoPrimaryKey.uid)
       })
 
@@ -57,9 +52,13 @@ describe.each([
         })
     })
     test(`${permission} key: create with primary key`, async () => {
-      await client.createIndex(uidAndPrimaryKey).then((response) => {
-        expect(response).toHaveProperty('uid', uidAndPrimaryKey.uid)
-      })
+      await client
+        .createIndex(uidAndPrimaryKey.uid, {
+          primaryKey: uidAndPrimaryKey.primaryKey,
+        })
+        .then((response) => {
+          expect(response).toHaveProperty('uid', uidAndPrimaryKey.uid)
+        })
       await client
         .getIndex(uidAndPrimaryKey.uid)
         .show()
@@ -123,7 +122,14 @@ describe.each([
       })
       await expect(client.listIndexes()).resolves.toHaveLength(1)
     })
-
+    test(`${permission} key: bad host should raise CommunicationError`, async () => {
+      const client = new MeiliSearch({ host: 'badHost' })
+      try {
+        await client.version()
+      } catch (e) {
+        expect(e.type).toEqual('MeiliSearchCommunicationError')
+      }
+    })
     test(`${permission} key: show deleted index should fail`, async () => {
       const index = client.getIndex(uidNoPrimaryKey.uid)
       await expect(index.show()).rejects.toThrowError(
@@ -132,14 +138,16 @@ describe.each([
     })
 
     test(`${permission} key: create index with already existing uid should fail`, async () => {
-      await expect(client.createIndex(uidAndPrimaryKey)).rejects.toThrowError(
-        `Impossible to create index; index already exists`
-      )
+      await expect(
+        client.createIndex(uidAndPrimaryKey.uid, {
+          primaryKey: uidAndPrimaryKey.primaryKey,
+        })
+      ).rejects.toThrowError(`index already exists`)
     })
 
     test(`${permission} key: create index with missing uid should fail`, async () => {
       // @ts-ignore
-      await expect(client.createIndex({ uid: null })).rejects.toThrowError(
+      await expect(client.createIndex(null)).rejects.toThrowError(
         `Index creation must have an uid`
       )
     })
@@ -157,116 +165,56 @@ describe.each([
         expect(response).toBe(true)
       })
     })
+    test(`${permission} key: get version`, async () => {
+      await client.version().then((response: Types.Version) => {
+        expect(response).toHaveProperty('commitSha', expect.any(String))
+        expect(response).toHaveProperty('buildDate', expect.any(String))
+        expect(response).toHaveProperty('pkgVersion', expect.any(String))
+      })
+    })
+    test(`${permission} key: get system info`, async () => {
+      await client.sysInfo().then((response: Types.SysInfo) => {
+        expect(response).toHaveProperty('memoryUsage', null)
+        expect(response).toHaveProperty('processorUsage', expect.any(Array))
+        expect(response.global).toHaveProperty(
+          'totalMemory',
+          expect.any(Number)
+        )
+        expect(response.global).toHaveProperty('usedMemory', expect.any(Number))
+        expect(response.global).toHaveProperty('totalSwap', expect.any(Number))
+        expect(response.global).toHaveProperty('usedSwap', expect.any(Number))
+        expect(response.global).toHaveProperty('inputData', expect.any(Number))
+        expect(response.global).toHaveProperty('outputData', expect.any(Number))
+        expect(response.process).toHaveProperty('memory', expect.any(Number))
+        expect(response.process).toHaveProperty('cpu', expect.any(Number))
+      })
+    })
+    test(`${permission} key: get pretty system info`, async () => {
+      await client.prettySysInfo().then((response: Types.SysInfoPretty) => {
+        expect(response).toHaveProperty('memoryUsage', expect.any(String))
+        expect(response).toHaveProperty('processorUsage', expect.any(Array))
+        expect(response.global).toHaveProperty(
+          'totalMemory',
+          expect.any(String)
+        )
+        expect(response.global).toHaveProperty('usedMemory', expect.any(String))
+        expect(response.global).toHaveProperty('totalSwap', expect.any(String))
+        expect(response.global).toHaveProperty('usedSwap', expect.any(String))
+        expect(response.global).toHaveProperty('inputData', expect.any(String))
+        expect(response.global).toHaveProperty('outputData', expect.any(String))
+        expect(response.process).toHaveProperty('memory', expect.any(String))
+        expect(response.process).toHaveProperty('cpu', expect.any(String))
+      })
+    })
+    test(`${permission} key: get /stats information`, async () => {
+      await client.stats().then((response: Types.Stats) => {
+        expect(response).toHaveProperty('databaseSize', expect.any(Number))
+        expect(response).toHaveProperty('lastUpdate') // TODO: Could be null, find out why
+        expect(response).toHaveProperty('indexes', expect.any(Object))
+      })
+    })
   })
 })
-
-describe.each([{ client: masterClient, permission: 'Master' }])(
-  'Test on routes where only master key has access',
-  ({ client, permission }) => {
-    describe('Test on base routes', () => {
-      test(`${permission} key: get version`, async () => {
-        await client.version().then((response: Types.Version) => {
-          expect(response).toHaveProperty('commitSha', expect.any(String))
-          expect(response).toHaveProperty('buildDate', expect.any(String))
-          expect(response).toHaveProperty('pkgVersion', expect.any(String))
-        })
-      })
-      test(`${permission} key: get system info`, async () => {
-        await client.sysInfo().then((response: Types.SysInfo) => {
-          expect(response).toHaveProperty('memoryUsage', null)
-          expect(response).toHaveProperty('processorUsage', expect.any(Array))
-          expect(response.global).toHaveProperty(
-            'totalMemory',
-            expect.any(Number)
-          )
-          expect(response.global).toHaveProperty(
-            'usedMemory',
-            expect.any(Number)
-          )
-          expect(response.global).toHaveProperty(
-            'totalSwap',
-            expect.any(Number)
-          )
-          expect(response.global).toHaveProperty('usedSwap', expect.any(Number))
-          expect(response.global).toHaveProperty(
-            'inputData',
-            expect.any(Number)
-          )
-          expect(response.global).toHaveProperty(
-            'outputData',
-            expect.any(Number)
-          )
-          expect(response.process).toHaveProperty('memory', expect.any(Number))
-          expect(response.process).toHaveProperty('cpu', expect.any(Number))
-        })
-      })
-      test(`${permission} key: get pretty system info`, async () => {
-        await client.prettySysInfo().then((response: Types.SysInfoPretty) => {
-          expect(response).toHaveProperty('memoryUsage', expect.any(String))
-          expect(response).toHaveProperty('processorUsage', expect.any(Array))
-          expect(response.global).toHaveProperty(
-            'totalMemory',
-            expect.any(String)
-          )
-          expect(response.global).toHaveProperty(
-            'usedMemory',
-            expect.any(String)
-          )
-          expect(response.global).toHaveProperty(
-            'totalSwap',
-            expect.any(String)
-          )
-          expect(response.global).toHaveProperty('usedSwap', expect.any(String))
-          expect(response.global).toHaveProperty(
-            'inputData',
-            expect.any(String)
-          )
-          expect(response.global).toHaveProperty(
-            'outputData',
-            expect.any(String)
-          )
-          expect(response.process).toHaveProperty('memory', expect.any(String))
-          expect(response.process).toHaveProperty('cpu', expect.any(String))
-        })
-      })
-      test(`${permission} key: get /stats information`, async () => {
-        await client.stats().then((response: Types.Stats) => {
-          expect(response).toHaveProperty('databaseSize', expect.any(Number))
-          expect(response).toHaveProperty('lastUpdate') // TODO: Could be null, find out why
-          expect(response).toHaveProperty('indexes', expect.any(Object))
-        })
-      })
-    })
-  }
-)
-
-describe.each([{ client: privateClient, permission: 'Private' }])(
-  'Test on routes where private key should not have access',
-  ({ client, permission }) => {
-    describe('Test on base routes', () => {
-      test(`${permission} key: try to get version and be denied`, async () => {
-        await expect(client.version()).rejects.toThrowError(
-          `Invalid API key: ${PRIVATE_KEY}`
-        )
-      })
-      test(`${permission} key: try to get system info and be denied`, async () => {
-        await expect(client.sysInfo()).rejects.toThrowError(
-          `Invalid API key: ${PRIVATE_KEY}`
-        )
-      })
-      test(`${permission} key: try to get pretty system info and be denied`, async () => {
-        await expect(client.prettySysInfo()).rejects.toThrowError(
-          `Invalid API key: ${PRIVATE_KEY}`
-        )
-      })
-      test(`${permission} key: try to get /stats information and be denied`, async () => {
-        await expect(client.stats()).rejects.toThrowError(
-          `Invalid API key: ${PRIVATE_KEY}`
-        )
-      })
-    })
-  }
-)
 
 describe.each([{ client: publicClient, permission: 'Public' }])(
   'Test on routes where public key should not have access',
@@ -278,14 +226,16 @@ describe.each([{ client: publicClient, permission: 'Public' }])(
         )
       })
       test(`${permission} key: try to create Index with primary key and be denied`, async () => {
-        await expect(client.createIndex(uidAndPrimaryKey)).rejects.toThrowError(
-          `Invalid API key: ${PUBLIC_KEY}`
-        )
+        await expect(
+          client.createIndex(uidAndPrimaryKey.uid, {
+            primaryKey: uidAndPrimaryKey.primaryKey,
+          })
+        ).rejects.toThrowError(`Invalid API key: ${PUBLIC_KEY}`)
       })
       test(`${permission} key: try to create Index with NO primary key and be denied`, async () => {
-        await expect(client.createIndex(uidNoPrimaryKey)).rejects.toThrowError(
-          `Invalid API key: ${PUBLIC_KEY}`
-        )
+        await expect(
+          client.createIndex(uidNoPrimaryKey.uid)
+        ).rejects.toThrowError(`Invalid API key: ${PUBLIC_KEY}`)
       })
       test(`${permission} key: try to get index info and be denied`, async () => {
         await expect(
@@ -299,7 +249,9 @@ describe.each([{ client: publicClient, permission: 'Public' }])(
       })
       test(`${permission} key: try to update index and be denied`, async () => {
         await expect(
-          client.getIndex(uidAndPrimaryKey.uid).updateIndex(uidAndPrimaryKey)
+          client
+            .getIndex(uidAndPrimaryKey.uid)
+            .updateIndex({ primaryKey: uidAndPrimaryKey.primaryKey })
         ).rejects.toThrowError(`Invalid API key: ${PUBLIC_KEY}`)
       })
     })
@@ -334,54 +286,58 @@ describe.each([{ client: anonymousClient, permission: 'No' }])(
     describe('Test on indexes', () => {
       test(`${permission} key: try to get all indexes and be denied`, async () => {
         await expect(client.listIndexes()).rejects.toThrowError(
-          `Invalid API key: Need a token`
+          `You must have an authorization token`
         )
       })
       test(`${permission} key: try to create an index with primary key and be denied`, async () => {
-        await expect(client.createIndex(uidAndPrimaryKey)).rejects.toThrowError(
-          `Invalid API key: Need a token`
-        )
+        await expect(
+          client.createIndex(uidAndPrimaryKey.uid, {
+            primaryKey: uidAndPrimaryKey.primaryKey,
+          })
+        ).rejects.toThrowError(`You must have an authorization token`)
       })
       test(`${permission} key: try to create an index with NO primary key and be denied`, async () => {
-        await expect(client.createIndex(uidNoPrimaryKey)).rejects.toThrowError(
-          `Invalid API key: Need a token`
-        )
+        await expect(
+          client.createIndex(uidNoPrimaryKey.uid)
+        ).rejects.toThrowError(`You must have an authorization token`)
       })
       test(`${permission} key: try to get index info and be denied`, async () => {
         await expect(
           client.getIndex(uidNoPrimaryKey.uid).show()
-        ).rejects.toThrowError(`Invalid API key: Need a token`)
+        ).rejects.toThrowError(`You must have an authorization token`)
       })
       test(`${permission} key: try to delete index and be denied`, async () => {
         await expect(
           client.getIndex(uidAndPrimaryKey.uid).deleteIndex()
-        ).rejects.toThrowError(`Invalid API key: Need a token`)
+        ).rejects.toThrowError(`You must have an authorization token`)
       })
       test(`${permission} key: try to update index and be denied`, async () => {
         await expect(
-          client.getIndex(uidAndPrimaryKey.uid).updateIndex(uidAndPrimaryKey)
-        ).rejects.toThrowError(`Invalid API key: Need a token`)
+          client
+            .getIndex(uidAndPrimaryKey.uid)
+            .updateIndex({ primaryKey: uidAndPrimaryKey.primaryKey })
+        ).rejects.toThrowError(`You must have an authorization token`)
       })
     })
     describe('Test on base routes', () => {
       test(`${permission} key: try to get version and be denied`, async () => {
         await expect(client.version()).rejects.toThrowError(
-          `Invalid API key: Need a token`
+          `You must have an authorization token`
         )
       })
       test(`${permission} key: try to get system info and be denied`, async () => {
         await expect(client.sysInfo()).rejects.toThrowError(
-          `Invalid API key: Need a token`
+          `You must have an authorization token`
         )
       })
       test(`${permission} key: try to get pretty system info and be denied`, async () => {
         await expect(client.prettySysInfo()).rejects.toThrowError(
-          `Invalid API key: Need a token`
+          `You must have an authorization token`
         )
       })
       test(`${permission} key: try to get /stats information and be denied`, async () => {
         await expect(client.stats()).rejects.toThrowError(
-          `Invalid API key: Need a token`
+          `You must have an authorization token`
         )
       })
     })